feat : Components existing on one page
