+++ app/_layout.tsx
@@ -4,6 +4,7 @@ import { Stack } from "expo-router";
 import { DefaultTheme, ThemeProvider } from "@react-navigation/native";
 import * as SplashScreen from "expo-splash-screen";
 import { StatusBar } from "expo-status-bar";
+import { TrainingDataProvider } from "@/components/training_management/training_data_context";
 
 export default function Layout() {
   const router = useRouter();
@@ -17,9 +18,13 @@ export default function Layout() {
 
   return (
     <ThemeProvider value={DefaultTheme}>
+    <TrainingDataProvider>
+
       <Stack>
         <Stack.Screen name="dashboard" options={{ headerShown: false }} />
       </Stack>
+    </TrainingDataProvider>
+
       <StatusBar style="auto" />
     </ThemeProvider>
   );
