+++ app/dashboard/training_management/index.tsx
@@ -1,5 +1,5 @@
 import TrainingManagementPage from "@/src/training_management_page/training_management_page";
-import UserManagementPage from "@/src/user_anagement_page/user_anagement_page";
+import UserManagementPage from "@/src/user_management_page/user_management_page";
 import React from "react";
 import { View, StyleSheet } from "react-native";
