+++ app/dashboard/training_stats/index.tsx
@@ -1,6 +1,6 @@
-import TrainingStats from "@/components/training_stats/training_stats";
+import TrainingStats from "@/components/training_stats/preference_circular_chart";
 import TrainingStatsPage from "@/src/training_stats_page/training_stats_page";
-import UserManagementPage from "@/src/user_anagement_page/user_anagement_page";
+import UserManagementPage from "@/src/user_management_page/user_management_page";
 import React from "react";
 import { View, StyleSheet } from "react-native";
