+++ app/dashboard/user_management/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import UserManagementPage from "@/src/user_anagement_page/user_anagement_page";
+import UserManagementPage from "@/src/user_management_page/user_management_page";
 
 export default function UserManagementIndex() {
   return <UserManagementPage />;
