+++ src/training_stats_page/training_stats_page.tsx
@@ -1,81 +1,58 @@
-import TrainingStats from "@/components/training_stats/training_stats";
+import PopularCardList from "@/components/training_stats/popular_card_list";
+import PreferenceCircularChart from "@/components/training_stats/preference_circular_chart";
 import React from "react";
-
-const popularGames = [
-  { name: "Popular Game 1", score: 95 },
-  { name: "Popular Game 2", score: 90 },
-  { name: "Popular Game 3", score: 85 },
-  { name: "Popular Game 4", score: 80 },
-];
-
-const unpopularGames = [
-  { name: "Unpopular Game 1", score: 40 },
-  { name: "Unpopular Game 2", score: 35 },
-  { name: "Unpopular Game 3", score: 30 },
-  { name: "Unpopular Game 4", score: 25 },
-];
+import { View, Text, StyleSheet } from "react-native";
 
 export default function TrainingStatsPage() {
   return (
-    <div
-      style={{
-        display: "flex",
-        flexDirection: "row",
-        justifyContent: "space-between",
-        alignItems: "flex-start",
-        width: "100%",
-        height: "100%",
-        padding: "20px",
-        boxSizing: "border-box",
-      }}
-    >
-      {/* 왼쪽: TrainingStats 호출 */}
-      <div style={{ flex: 1, marginRight: "20px" }}>
-        <TrainingStats />
-      </div>
+    <View style={styles.container}>
+      {/* 페이지 제목 */}
+      <Text style={styles.title}>Training Stats Page</Text>
 
-      {/* 오른쪽: 카드 레이아웃 */}
-      <div style={{ flex: 1 }}>
-        {/* 인기 있는 게임 */}
-        <h3>Popular Games</h3>
-        <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: "10px" }}>
-          {popularGames.map((game, index) => (
-            <div
-              key={index}
-              style={{
-                padding: "10px",
-                border: "1px solid #ccc",
-                borderRadius: "8px",
-                backgroundColor: "#f9f9f9",
-                textAlign: "center",
-              }}
-            >
-              <h4 style={{ margin: "10px 0" }}>{game.name}</h4>
-              <p>Score: {game.score}</p>
-            </div>
-          ))}
-        </div>
+      {/* 메인 콘텐츠 */}
+      <View style={styles.content}>
+        {/* 왼쪽: PreferenceCircularChart */}
+        <View style={styles.left}>
+          <PreferenceCircularChart />
+        </View>
 
-        {/* 비인기 게임 */}
-        <h3 style={{ marginTop: "20px" }}>Unpopular Games</h3>
-        <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: "10px" }}>
-          {unpopularGames.map((game, index) => (
-            <div
-              key={index}
-              style={{
-                padding: "10px",
-                border: "1px solid #ccc",
-                borderRadius: "8px",
-                backgroundColor: "#f9f9f9",
-                textAlign: "center",
-              }}
-            >
-              <h4 style={{ margin: "10px 0" }}>{game.name}</h4>
-              <p>Score: {game.score}</p>
-            </div>
-          ))}
-        </div>
-      </div>
-    </div>
+        {/* 오른쪽: PopularCardList */}
+        <View style={styles.right}>
+          <PopularCardList />
+        </View>
+      </View>
+    </View>
   );
 }
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    padding: 16,
+    backgroundColor: "#fff",
+  },
+  title: {
+    fontSize: 24,
+    fontWeight: "bold",
+    textAlign: "center",
+    marginBottom: 20,
+    color: "#333",
+  },
+  content: {
+    flex: 1,
+    flexDirection: "row",
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  left: {
+    flex: 1,
+    alignItems: "center",
+    justifyContent: "center",
+    marginRight: 16,
+  },
+  right: {
+    flex: 1,
+    alignItems: "center",
+    justifyContent: "center",
+  },
+});
